+++ pyfakefs/pytest_tests/pytest_check_failed_plugin_test.py
@@ -1,9 +1,14 @@
 """Tests that a failed pytest properly displays the call stack.
-Uses the output from running pytest with pytest_plugin_failing_test.py.
+Uses the output from running pytest with pytest_plugin_failing_helper.py.
 Regression test for #381.
 """
+import os
+
+import pytest
 
 
+@pytest.mark.skipif(not os.path.exists('testresult.txt'),
+                    reason='Only run in CI tests')
 def test_failed_testresult_stacktrace():
     with open('testresult.txt') as f:
         contents = f.read()